Fix silt bug with nested child observables
We should ignore components where `type` isn't a string (for now).
We'll include a prop later to lift the events into the stream. Also
break some pieces out into separate modules.
